Resolve interface names on control-interface too
