+++ nsd-control.c
@@ -495,6 +495,7 @@ go(const char* cfgfile, char* svr, int argc, char* argv[])
 	}
 	if(!opt->control_enable)
 		fprintf(stderr, "warning: control-enable is 'no' in the config file.\n");
+	resolve_interface_names(opt);
 	ctx = setup_ctx(opt);
 
 	/* contact server */
