+++ options.h
@@ -471,5 +471,9 @@ void config_apply_pattern(struct pattern_options *dest, const char* name);
  * when a fileread fails because it is a directory, helps the user figure
  * out what just happened */
 void warn_if_directory(const char* filetype, FILE* f, const char* fname);
+/* resolve interface names in the options "ip-address:" (or "interface:")
+ * and "control-interface:" into the ip-addresses associated with those
+ * names. */
+void resolve_interface_names(struct nsd_options* options);
 
 #endif /* OPTIONS_H */
